+++ packages/studio-base/src/context/PlayerSelectionContext.ts
@@ -50,9 +50,6 @@ export interface IDataSourceFactory {
   hidden?: boolean;
   warning?: string | JSX.Element;
 
-  /** Whether to wait for a user to be logged in before initializing this source. */
-  currentUserRequired?: boolean;
-
   sampleLayout?: LayoutData;
 
   formConfig?: {
